refactor: Split WorkoutSelectionModal into smaller components and custom hooks
- Refactored WorkoutSelectionModal into a more maintainable structure using custom hooks.
- Created five custom hooks: useWorkoutData, useWorkoutSelection, useWorkoutScheduling, useWorkoutEditing, and useModalState.
- Updated the main component to focus on rendering and UI logic.
- Improved testability and readability by separating concerns.
- Renamed onClose prop to onCloseAction for consistency with Next.js client component requirements.
- Added README documentation for the new hooks and refactored structure.
