+++ src/app/(admin)/admin/teams/[teamSlug]/_hooks/useWorkoutScheduling.ts
@@ -1,0 +1,218 @@
+import { useCallback } from "react"
+import { toast } from "sonner"
+import { useServerAction } from "zsa-react"
+import {
+	deleteScheduledWorkoutAction,
+	scheduleStandaloneWorkoutAction,
+	scheduleWorkoutAction,
+	updateScheduledWorkoutAction,
+} from "../_actions/scheduling-actions"
+import type {
+	ProgrammingTrack,
+	ScheduledWorkoutWithDetails,
+	StandaloneWorkout,
+	TrackWorkout,
+} from "../_components/workout-selection"
+
+interface UseWorkoutSchedulingProps {
+	teamId: string
+	selectedDate: Date | null
+	selectedTrack: ProgrammingTrack | null
+	selectedWorkout:
+		| (TrackWorkout & {
+				isScheduled?: boolean
+				lastScheduledAt?: Date | null
+		  })
+		| null
+	selectedStandaloneWorkout:
+		| (StandaloneWorkout & {
+				isScheduled?: boolean
+				lastScheduledAt?: Date | null
+		  })
+		| null
+	onWorkoutScheduledAction: () => void
+	onCloseAction: () => void
+	loadScheduledWorkouts: () => void
+}
+
+interface SchedulingFormData {
+	classTimes: string
+	teamNotes: string
+	scalingGuidance: string
+}
+
+export function useWorkoutScheduling({
+	teamId,
+	selectedDate,
+	selectedTrack,
+	selectedWorkout,
+	selectedStandaloneWorkout,
+	onWorkoutScheduledAction,
+	onCloseAction,
+	loadScheduledWorkouts,
+}: UseWorkoutSchedulingProps) {
+	// Server actions
+	const { execute: scheduleWorkout, isPending: isScheduling } = useServerAction(
+		scheduleWorkoutAction,
+	)
+	const {
+		execute: scheduleStandaloneWorkout,
+		isPending: isSchedulingStandalone,
+	} = useServerAction(scheduleStandaloneWorkoutAction)
+	const { execute: updateScheduledWorkout, isPending: isUpdatingScheduled } =
+		useServerAction(updateScheduledWorkoutAction)
+	const { execute: deleteScheduledWorkout, isPending: isDeletingScheduled } =
+		useServerAction(deleteScheduledWorkoutAction)
+
+	// Scheduling handlers
+	const handleScheduleWorkout = useCallback(
+		async (formData: SchedulingFormData) => {
+			const { classTimes, teamNotes, scalingGuidance } = formData
+
+			if (process.env.LOG_LEVEL === "info") {
+				console.log(
+					`INFO: [WorkoutSelectionModal] Scheduling workout: track=${
+						selectedTrack?.id
+					}, workout=${selectedWorkout?.id || selectedStandaloneWorkout?.id}`,
+				)
+			}
+
+			// Check if either a track workout or standalone workout is selected
+			if (!selectedWorkout && !selectedStandaloneWorkout) {
+				toast.error("Please select a workout")
+				return
+			}
+
+			if (!selectedDate) {
+				toast.error("Please select a date")
+				return
+			}
+
+			// Handle standalone workouts
+			if (selectedStandaloneWorkout) {
+				const [result] = await scheduleStandaloneWorkout({
+					teamId,
+					workoutId: selectedStandaloneWorkout.id,
+					scheduledDate: selectedDate.toISOString(),
+					teamSpecificNotes: teamNotes || undefined,
+					scalingGuidanceForDay: scalingGuidance || undefined,
+					classTimes: classTimes || undefined,
+				})
+
+				if (result?.success) {
+					console.log(
+						`INFO: [WorkoutScheduling] Workflow completed: selected standalone workout '${
+							selectedStandaloneWorkout.id
+						}' scheduled for '${
+							selectedDate.toISOString().split("T")[0]
+						}' with ${classTimes ? "1" : "0"} class time`,
+					)
+					toast.success("Standalone workout scheduled successfully!")
+					onWorkoutScheduledAction()
+					onCloseAction()
+				} else {
+					toast.error("Failed to schedule standalone workout")
+				}
+				return
+			}
+
+			// Handle track workouts
+			if (selectedWorkout) {
+				const [result] = await scheduleWorkout({
+					teamId,
+					trackWorkoutId: selectedWorkout.id,
+					scheduledDate: selectedDate.toISOString(),
+					teamSpecificNotes: teamNotes || undefined,
+					scalingGuidanceForDay: scalingGuidance || undefined,
+					classTimes: classTimes || undefined,
+				})
+
+				if (result?.success) {
+					console.log(
+						`INFO: [WorkoutScheduling] Workflow completed: selected workout '${
+							selectedWorkout.workoutId
+						}' from track '${selectedTrack?.id}' scheduled for '${
+							selectedDate.toISOString().split("T")[0]
+						}' with ${classTimes ? "1" : "0"} class time`,
+					)
+					toast.success("Workout scheduled successfully!")
+					onWorkoutScheduledAction()
+					onCloseAction()
+				} else {
+					toast.error("Failed to schedule workout")
+				}
+			}
+		},
+		[
+			selectedWorkout,
+			selectedStandaloneWorkout,
+			selectedDate,
+			selectedTrack,
+			teamId,
+			scheduleWorkout,
+			scheduleStandaloneWorkout,
+			onWorkoutScheduledAction,
+			onCloseAction,
+		],
+	)
+
+	// Scheduled workout management handlers
+	const handleUpdateScheduled = useCallback(
+		async (instanceId: string, formData: SchedulingFormData) => {
+			const { classTimes, teamNotes, scalingGuidance } = formData
+
+			const [result] = await updateScheduledWorkout({
+				instanceId,
+				teamSpecificNotes: teamNotes || undefined,
+				scalingGuidanceForDay: scalingGuidance || undefined,
+				classTimes: classTimes || undefined,
+			})
+
+			if (result?.success) {
+				toast.success("Scheduled workout updated successfully!")
+				loadScheduledWorkouts() // Reload the list
+				onWorkoutScheduledAction() // Refresh the calendar
+				return true
+			}
+			toast.error("Failed to update scheduled workout")
+			return false
+		},
+		[updateScheduledWorkout, loadScheduledWorkouts, onWorkoutScheduledAction],
+	)
+
+	const handleDeleteScheduled = useCallback(
+		async (instanceId: string) => {
+			const [result] = await deleteScheduledWorkout({
+				instanceId,
+				teamId,
+			})
+
+			if (result?.success) {
+				toast.success("Scheduled workout removed successfully!")
+				loadScheduledWorkouts() // Reload the list
+				onWorkoutScheduledAction() // Refresh the calendar
+				return true
+			}
+			toast.error("Failed to remove scheduled workout")
+			return false
+		},
+		[
+			deleteScheduledWorkout,
+			teamId,
+			loadScheduledWorkouts,
+			onWorkoutScheduledAction,
+		],
+	)
+
+	return {
+		// Handlers
+		handleScheduleWorkout,
+		handleUpdateScheduled,
+		handleDeleteScheduled,
+		// Loading states
+		isScheduling,
+		isSchedulingStandalone,
+		isUpdatingScheduled,
+		isDeletingScheduled,
+	}
+}
